requirements.txt
flask
gunicorn
google-generativeai
textblob
pysqlite3